+++ lib/recurly/api.rb
@@ -16,6 +16,8 @@ module Recurly
 
     @@base_uri = "https://api.recurly.com/v2/"
 
+    RECURLY_API_VERSION = '2.1'
+
     FORMATS = Helper.hash_with_indifferent_read_access(
       'pdf' => 'application/pdf',
       'xml' => 'application/xml'
@@ -25,7 +27,7 @@ module Recurly
       # Additional HTTP headers sent with each API call
       # @return [Hash{String => String}]
       def headers
-        @headers ||= { 'Accept' => accept, 'User-Agent' => user_agent }
+        @headers ||= { 'Accept' => accept, 'User-Agent' => user_agent, 'X-Api-Version' => RECURLY_API_VERSION }
       end
 
       # @return [String, nil] Accept-Language header value