+++ lib/recurly/coupon.rb
@@ -24,11 +24,14 @@ module Recurly
       temporal_unit
       temporal_amount
       max_redemptions
+      max_redemptions_per_account
       applies_to_all_plans
       created_at
       plan_codes
       description
       invoice_description
+      applies_to_non_plan_charges
+      redemption_resource
     )
     alias to_param coupon_code
 
@@ -46,14 +49,16 @@ module Recurly
     #
     # @return [true]
     # @param account_or_code [Account, String]
+    # @param currency [String]
+    # @param opts [Hash]
     # @example
     #   coupon = Coupon.find coupon_code
-    #   coupon.redeem account_code
+    #   coupon.redeem account_code, 'USD', subscription_uuid: 'ab3b1dbabc3195'
     #
     #   coupon = Coupon.find coupon_code
     #   account = Account.find account_code
     #   coupon.redeem account
-    def redeem account_or_code, currency = nil
+    def redeem account_or_code, currency = nil, extra_opts={}
       return false unless link? :redeem
 
       account_code = if account_or_code.is_a? Account
@@ -62,11 +67,15 @@ module Recurly
         account_or_code
       end
 
+      redemption_options = {
+        :account_code => account_code,
+        :currency     => currency || Recurly.default_currency
+      }.merge(extra_opts)
+
+      redemption = redemptions.new(redemption_options)
+
       Redemption.from_response follow_link(:redeem,
-        :body => (redemption = redemptions.new(
-          :account_code => account_code,
-          :currency     => currency || Recurly.default_currency
-        )).to_xml
+        :body => redemption.to_xml
       )
     rescue API::UnprocessableEntity => e
       redemption.apply_errors e
@@ -77,6 +86,10 @@ module Recurly
       redemption = redeem account_code, currency
       raise Invalid.new(self) unless redemption.persisted?
       redemption
+    end
+
+    def unlimited_redemptions_per_account?
+      !max_redemptions_per_account
     end
   end
 end